+++ src/views/alerts.js
@@ -4,9 +4,9 @@ import usage from './alerts.usage.js';
 
 export default function(discovery) {
     function render(el, config, data, context) {
-        const { content } = config;
+        const { content = 'text' } = config;
 
-        discovery.view.render(el, content || 'text', data, context);
+        return discovery.view.render(el, content, data, context);
     }
 
     discovery.view.define('alert', render, { usage });