+++ src/views/lists.js
@@ -3,7 +3,7 @@ import usage from './lists.usage.js';
 
 export default function(discovery) {
     function render(el, config, data, context) {
-        const { item = 'text', itemConfig, limit, emptyText } = config;
+        const { item, itemConfig, limit, emptyText } = config;
 
         if (emptyText !== false && emptyText !== '') {
             el.setAttribute('emptyText', emptyText || 'Empty list');
@@ -14,7 +14,7 @@ export default function(discovery) {
         }
 
         if (Array.isArray(data)) {
-            discovery.view.renderList(el, this.composeConfig({
+            return discovery.view.renderList(el, this.composeConfig({
                 view: 'list-item',
                 content: item
             }, itemConfig), data, context, 0, discovery.view.listLimit(limit, 25));