+++ src/views/menu-item.js
@@ -3,12 +3,16 @@
 export default function(discovery) {
     discovery.view.define('menu-item', function(el, config, data, context) {
         const { content, onClick } = config;
-        const { text, selected = false, disabled = false } = data;
+        const { text, selected = false, disabled = false, href, external } = data || {};
 
         if (disabled) {
             el.classList.add('disabled');
         } else if (typeof onClick === 'function') {
-            el.addEventListener('click', () => onClick(data));
+            el.addEventListener('click', () => onClick(data, context));
+            el.classList.add('onclick');
+        } else if (href) {
+            el.href = href;
+            el.target = external ? '_blank' : '';
         }
 
         if (selected) {
@@ -16,9 +20,11 @@ export default function(discovery) {
         }
 
         if (content) {
-            discovery.view.render(el, content, data, context);
+            return discovery.view.render(el, content, data, context);
         } else {
-            el.innerText = typeof data === 'string' ? data : text || 'Untitled item';
+            el.textContent = typeof data === 'string' ? data : text || 'Untitled item';
         }
+    }, {
+        tag: 'a'
     });
 }