+++ src/views/nav-button.js
@@ -1,16 +1,16 @@
 /* eslint-env browser */
 
-import usage from './button.usage.js';
-
 export default function(discovery) {
-    function render(el, config, data, context) {
-        const { content, disabled = false, onClick } = config;
+    discovery.view.define('nav-button', function(el, config, data, context) {
+        const { name, content, disabled = false, onClick } = config;
         const { text = '', href, external } = data || {};
 
-        el.classList.add('view-button');
+        if (name) {
+            el.dataset.name = name;
+        }
 
         if (discovery.query(disabled, data, context)) {
-            el.disabled = true;
+            el.classList.add('disabled');
         } else if (typeof onClick === 'function') {
             el.addEventListener('click', () => onClick(el, data, context));
             el.classList.add('onclick');
@@ -24,10 +24,7 @@ export default function(discovery) {
         } else {
             el.textContent = text;
         }
-    }
-
-    discovery.view.define('button', render, { tag: 'button', usage });
-    discovery.view.define('button-primary', render, { tag: 'button', usage });
-    discovery.view.define('button-danger', render, { tag: 'button', usage });
-    discovery.view.define('button-warning', render, { tag: 'button', usage });
+    }, {
+        tag: 'a'
+    });
 }